# SmartVillageAPI/publish/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Smart Village API",
    "description": "API for Smart Village Management System",
    "version": "v1"
  },
  "paths": {
...
    "/api/Devices/Control/{id}": {
      "post": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Irrigation/Dashboard/{userId}": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zones/{userId}": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone/{id}": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone": {
      "post": {
        "tags": [
          "Irrigation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIrrigationZoneDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIrrigationZoneDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIrrigationZoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone/{id}/Settings": {
      "put": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIrrigationSettingsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIrrigationSettingsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIrrigationSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone/{id}/Control": {
      "post": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlIrrigationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlIrrigationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ControlIrrigationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone/{id}/SoilReading": {
      "post": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoilReadingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SoilReadingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SoilReadingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Logs/{userId}": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/AiRecommendation/{userId}": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Irrigation/Zone/{id}/AiRecommendation": {
      "get": {
        "tags": [
          "Irrigation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ControlDeviceDto": {
        "type": "object",
        "properties": {
          "newState": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ControlIrrigationDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateIrrigationZoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "zoneId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "moistureThreshold": {
            "type": "number",
            "format": "double"
          },
          "plantType": {
            "$ref": "#/components/schemas/PlantType"
          }
        },
        "additionalProperties": false
      },
...
      "PlantType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13,
          14,
          15,
          16,
          17,
          18
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SoilReadingDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "UpdateIrrigationSettingsDto": {
        "type": "object",
        "properties": {
          "plantType": {
            "$ref": "#/components/schemas/PlantType"
          },
          "moistureThreshold": {
            "type": "number",
            "format": "double"
          },
          "isAutoMode": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token in the text input below.\r\n\r\nExample: \"12345abcdef\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}